Add authentication controller to handle user login and returning new valid token

+++ controllers/bookingController.js
@@ -9,24 +9,27 @@ let formatBooking = booking => {
     };
 }
 
-let bookings = async () => {
-    const allBookings = await Booking.find({});
+let bookings = async (args, req) => {
+    if(!req.isAuth) throw new Error('User is not authenticated!');
+    const allBookings = await Booking.find({user: req.user._id});
     return allBookings.map(booking => {
         return formatBooking(booking);
     })
 }
 
-let addBooking = async ({eventId}) => {
+let addBooking = async ({eventId}, req) => {
+    if(!req.isAuth) throw new Error('User is not authenticated!');
     const event = await Event.findById(eventId);
     const booking = new Booking({
-        user: '5f41949ec52cdd3a1b5c801b',
+        user: req.user._id,
         event: event._doc._id
     });
     const savedBooking = await booking.save();
     return formatBooking(savedBooking);
 };
 
-let cancelBooking = async ({bookingId}) => {
+let cancelBooking = async ({bookingId}, req) => {
+    if(!req.isAuth) throw new Error('User is not authenticated!');
     const booking = await Booking.findById(bookingId);
     await Booking.deleteOne({_id: booking._doc._id});
     return {
